+++ cmd/cli/get.go
@@ -43,12 +43,16 @@ and eventually URL to access Token view and Expert analysis view.`,
 			return
 		}
 
-		disableSSLChecking, err := cmd.Flags().GetBool("insecure")
+		insecure, err := cmd.Flags().GetBool("insecure")
 		if err != nil {
 			return
 		}
 
-		client, err := gdetect.NewClient(apiEndpoint, apiToken, disableSSLChecking, nil)
+		client, err := gdetect.NewClientFromConfig(gdetect.ClientConfig{
+			Endpoint: apiEndpoint,
+			Token:    apiToken,
+			Insecure: insecure,
+		})
 		if err != nil {
 			return
 		}