+++ cmd/cli/submit.go
@@ -45,7 +45,7 @@ analysis UUID and print it out.`,
 			return
 		}
 
-		disableSSLChecking, err := cmd.Flags().GetBool("insecure")
+		insecure, err := cmd.Flags().GetBool("insecure")
 		if err != nil {
 			return
 		}
@@ -70,7 +70,11 @@ analysis UUID and print it out.`,
 			return
 		}
 
-		client, err := gdetect.NewClient(apiEndpoint, apiToken, disableSSLChecking, nil)
+		client, err := gdetect.NewClientFromConfig(gdetect.ClientConfig{
+			Endpoint: apiEndpoint,
+			Token:    apiToken,
+			Insecure: insecure,
+		})
 		if err != nil {
 			return
 		}